fix warnings

+++ src/fixed-contact-ekf-flex-estimator-imu.cpp
@@ -85,7 +85,7 @@ namespace flexibilityEstimation
 
     void FixedContactEKFFlexEstimatorIMU::setMeasurement(const Vector & y)
     {
-        BOOST_ASSERT((getMeasurementSize()==y.size()) &&
+        BOOST_ASSERT((getMeasurementSize()==unsigned(y.size())) &&
                 "ERROR: The measurement vector has incorrect size");
 
 
@@ -142,7 +142,7 @@ namespace flexibilityEstimation
     void FixedContactEKFFlexEstimatorIMU::setMeasurementNoiseCovariance
                                             (const Matrix & R)
     {
-        BOOST_ASSERT(R.rows()==getMeasurementSize() &&
+        BOOST_ASSERT(unsigned(R.rows())==getMeasurementSize() &&
                      R.cols()==getMeasurementSize() &&
                     "ERROR: The measurement noise covariance matrix R has \
                         incorrect size");
@@ -175,7 +175,7 @@ namespace flexibilityEstimation
         Matrix R = ekf_.getRmatrixIdentity();
         R.block(0,0, getMeasurementSize() , getMeasurementSize())=R_;
 
-        for (int i= getMeasurementSize() ; i<ekf_.getMeasureSize() ; ++i)
+        for (unsigned i= getMeasurementSize() ; i<ekf_.getMeasureSize() ; ++i)
         {
             R(i,i)=virtualMeasurementCovariance_;
         }
